Split Overlay into two parts
-OverlayBottom
-OverlayQuality
